pyproject.toml
[project]
name = "django-rubble"
dynamic = ["version", "classifiers"]
description = "Extend django-model-utils and django-extensions."
readme = "README.md"
license = "MIT"
authors = [{ name = "Karl Wooster", email = "karl@woostertech.com" }]
keywords = ["django", "django-model-utils", "django-extensions"]
requires-python = ">=3.11,<4.0"
dependencies = [
  "django (>4,<5.3)",
  "case-converter (>=1.2.0,<2.0.0)",
  "toml (>=0.10.2,<0.11.0)",
  "pydantic (>=2.7)",
  "typing-extensions (>=4.5) ; python_version < '3.12'",
  "django-model-utils (>=4.5)",
  "loguru (>=0.7)",
  "croniter (>=3.0)",
  "pydantic-settings (>=2.7)",
  "attrmagic @ git+https://github.com/WoosterTech/AttrMagic.git",
  "furl (>=2.1.3,<3.0.0)",
  "levenshtein (>=0.26.1,<0.27.0)",
]

[project.urls]
source = "https://github.com/WoosterTech/django-rubble/"
changelog = "https://github.com/WoosterTech/django-rubble/blob/main/CHANGELOG.md"
issues = "https://github.com/WoosterTech/django-rubble/issues"
documentation = "https://woostertech.github.io/django-rubble/"

[project.optional-dependencies]
neapolitan = ["neapolitan (>=24.6)"]
history = ["django-simple-history (>=3.7)"]
secrets = ["infisical-python (>=2.3.5)"]
cli = ["typer (>=0.12)", "rich (>=13.7)"]

all = ["neapolitan (>=24.6)", "django-simple-history (>=3.7)", "infisical-python (>=2.3.5)", "typer (>=0.12)", "rich (>=13.7)"]

[project.scripts]
rubble = "django_rubble.cli:app"

[tool.poetry]
version = "0.7.0b6"
classifiers = [
  "Framework :: Django :: 4",
  "Framework :: Django :: 4.0",
  "Framework :: Django :: 4.1",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.0",
  "Framework :: Django :: 5.1",
]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
commitizen = "^3.27.0"
types-toml = "^0.10.8.20240310"
deptry = "^0.23.0"
basedpyright = "^1.33.0"
django-types = "^0.22.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
pytest-sugar = "^1.0.0"
pytest-django = "^4.8.0"
factory-boy = "^3.3.1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocstrings = {extras = ["python"], version = "^0.25.2"}
mkdocs-material = "^9.5.33"
mkdocs-gen-files = "^0.5.0"
mkdocs-literate-nav = "^0.6.1"
mkdocs-section-index = "^0.3.9"
mkdocs-include-markdown-plugin = "^6.2.2"
griffe-fieldz = "^0.2.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "test.settings"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.deptry]
extend_exclude = [".*/test_.*.py", ".*/factories.py"]

[tool.deptry.per_rule_ignores]
DEP004 = ["mkdocs_gen_files"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
